fmt, address review
