+++ prusti-viper/src/encoder/mir/contracts/mod.rs
@@ -1,8 +1,9 @@
 //! Procedure contracts enriched with borrow information.
 
 mod borrows;
+#[allow(clippy::module_inception)]
+mod contracts;
 mod interface;
 
-pub(crate) use interface::{
-    ContractsEncoderInterface, ContractsEncoderState, ProcedureContract, ProcedureContractMirDef,
-};
+pub(crate) use contracts::{ProcedureContract, ProcedureContractMirDef};
+pub(crate) use interface::{ContractsEncoderInterface, ContractsEncoderState};
