+++ prusti-viper/src/encoder/mir/pure/pure_functions/interpreter.rs
@@ -10,8 +10,8 @@ use crate::encoder::{
         generics::HighGenericsEncoderInterface, types::HighTypeEncoderInterface,
     },
     mir::{
-        sequences::MirSequencesEncoderInterface,
         pure::{specifications::SpecificationEncoderInterface, PureEncodingContext},
+        sequences::MirSequencesEncoderInterface,
         specifications::SpecificationsInterface,
         types::MirTypeEncoderInterface,
     },
@@ -20,7 +20,6 @@ use crate::encoder::{
     snapshot::interface::SnapshotEncoderInterface,
     Encoder,
 };
-
 use log::{debug, trace};
 use prusti_common::vir_local;
 use rustc_hash::FxHashMap;
